The whole thing is now Tampermonkey
