+++ README.md
@@ -1,30 +1,24 @@
 # FicTrail
 
-A Firefox extension for viewing and searching your AO3 history. Currently in alpha (rawr), so expect some rough edges.
+A Tampermonkey userscript for viewing and searching your AO3 history. Currently in alpha (rawr), so expect some rough edges.
 
 Supports search by title, author, fandom, and summary. Full-text search is on the way!
 
 Built with flaming love for the AO3 community.
 
-## Installation (Manual)
+## Installation
 
-Since this extension isn't on the Mozilla Add-ons store yet, you'll need to install it manually.
+1. **Install Tampermonkey**
+   - Install the [Tampermonkey browser extension](https://www.tampermonkey.net/) for your browser
 
-1. **Download the Extension**
-   - **Recommended**: [Download the latest release ZIP](https://github.com/serpentineegg/fictrail/releases/latest) (pre-built and ready to install)
-   - **Alternative**: [Download source ZIP](https://github.com/serpentineegg/fictrail/archive/refs/heads/main.zip) (development version)
-   - Extract the files to a folder on your computer
+2. **Install FicTrail**
+   - [Download the latest release](https://github.com/serpentineegg/fictrail/releases/latest) and click on `fictrail.user.js`
+   - Tampermonkey will automatically detect the userscript and prompt you to install it
 
-2. **Load the Extension**
-   - Open Firefox
-   - Type `about:debugging` in the address bar and press Enter
-   - Click "This Firefox" in the left sidebar
-   - Click "Load Temporary Add-on..."
-   - Navigate to the extension folder and select `manifest.json`
-
-3. **Pin the Extension** (Optional)
-   - Click the puzzle piece icon in your toolbar
-   - Find "FicTrail" and click the pin icon to keep it visible
+3. **Start Using FicTrail**
+   - Visit any AO3 page while logged in
+   - Look for the floating "FicTrail" button
+   - Click it to open your reading history and start searching!
 
 ## Privacy
 
@@ -32,6 +26,6 @@ All data processing happens locally in your browser. We don't collect, store, or
 
 ## Support
 
-Love this extension? Consider:
+Love this userscript? Consider:
 - 🐛 [Reporting bugs or suggesting features via GitHub Issues](https://github.com/serpentineegg/fictrail/issues)
 - ☕ [Buying me the meanest dark roast](https://ko-fi.com/serpentineegg)
