--- history.html
@@ -1,117 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>FicTrail - AO3 History Viewer</title>
-    <link rel="stylesheet" href="page-styles.css">
-</head>
-<body>
-<div class="container">
-    <header class="page-header">
-        <div class="header-content">
-            <div class="header-left">
-                <h1>📚 FicTrail</h1>
-                <p id="header-subtitle">Your AO3 History</p>
-            </div>
-            <div class="header-right">
-                <div class="header-actions">
-                    <div class="page-selector" id="page-selector" style="display: none;">
-                        <div class="page-selector-header">
-                            <label for="pages-slider">Pages to load</label>
-                            <span class="info-tooltip" title="Loading many pages can take a long time and use significant computer resources. Start with fewer pages for better performance.">ℹ️</span>
-                        </div>
-                        <div class="slider-container">
-                            <div class="slider-track">
-                                <span class="slider-min">1</span>
-                                <input type="range" id="pages-slider" min="1" max="50" value="10" class="slider">
-                                <span class="slider-max">50</span>
-                            </div>
-                        </div>
-                    </div>
-                    <button id="load-history-btn" class="primary-btn">Load My History</button>
-                </div>
-            </div>
-        </div>
-    </header>
-
-    <main class="main-content">
-        <div id="instructions" class="welcome-section">
-            <div class="welcome-content">
-                <h2>Ready to explore?</h2>
-                <p>Make sure you're logged into AO3 with history enabled, then hit the button above.</p>
-            </div>
-        </div>
-
-        <div id="loading" class="loading-section" style="display: none;">
-            <div class="loading-content">
-                <div class="spinner"></div>
-                <h2>Summoning your fic history...</h2>
-                <p id="loading-status">Diving deep into your AO3 rabbit hole...</p>
-                <p class="loading-note">This might take a hot minute (we're being respectful to AO3's servers) ✨</p>
-            </div>
-        </div>
-
-        <div id="error" class="error-section" style="display: none;">
-            <div class="error-content">
-                <h2>💀 Plot Twist!</h2>
-                <p id="error-message"></p>
-                <button id="navigate-btn" class="primary-btn" style="display: none;">Navigate</button>
-                <button id="retry-btn" class="secondary-btn">Try Again (Please?)</button>
-            </div>
-        </div>
-
-        <div id="history-container" class="history-section" style="display: none;">
-            <div class="history-header">
-                <div class="history-controls">
-                    <div class="search-section">
-                        <div class="search-instructions">Search by fandoms, titles, tags, authors, or summaries</div>
-                        <div class="search-box">
-                            <input type="text" id="search-input" placeholder="Enter your search...">
-                            <button id="search-btn">🔍</button>
-                        </div>
-                    </div>
-                    <div class="filter-controls">
-                        <select id="fandom-filter">
-                            <option value="">All Fandoms</option>
-                        </select>
-                    </div>
-                </div>
-            </div>
-
-
-            <div id="works-list" class="works-grid"></div>
-
-            <div id="no-results" class="no-results" style="display: none;">
-                <h3>404: Fics Not Found</h3>
-                <p>Your search came up empty! Try different keywords or maybe you haven't read that trope yet? 👀</p>
-            </div>
-        </div>
-    </main>
-
-    <footer class="page-footer">
-        <div class="footer-content">
-            <div class="footer-text">
-                <p>Made with ❤️‍🔥 by a baby fanboi</p>
-            </div>
-            <div class="footer-links">
-                <a href="https://ko-fi.com/serpentineegg" target="_blank" rel="noopener noreferrer" class="footer-link kofi-link">
-                    <span class="link-icon">☕</span>
-                    <span>Caffeinate me</span>
-                </a>
-                <a href="https://github.com/serpentineegg" target="_blank" rel="noopener noreferrer" class="footer-link">
-                    <span class="link-icon">🐙</span>
-                    <span>GitHub</span>
-                </a>
-                <a href="https://archiveofourown.org/users/painless" target="_blank" rel="noopener noreferrer" class="footer-link">
-                    <span class="link-icon">📖</span>
-                    <span>AO3</span>
-                </a>
-            </div>
-        </div>
-    </footer>
-</div>
-
-<script src="history.js"></script>
-</body>
-</html>
