--- manifest.json
@@ -1,37 +1,0 @@
-{
-  "manifest_version": 3,
-  "name": "FicTrail",
-  "version": "0.0.0-dev",
-  "description": "AO3 history search extension",
-  
-  "browser_specific_settings": {
-    "gecko": {
-      "id": "fictrail@serpentineegg.dev"
-    }
-  },
-
-  "permissions": [
-    "tabs"
-  ],
-
-  "host_permissions": [
-    "*://archiveofourown.org/*"
-  ],
-
-  "content_scripts": [
-    {
-      "matches": ["*://archiveofourown.org/*"],
-      "js": ["content.js"],
-      "run_at": "document_idle",
-      "all_frames": false
-    }
-  ],
-
-  "action": {
-    "default_title": "FicTrail - AO3 History Viewer"
-  },
-
-  "background": {
-    "scripts": ["background.js"]
-  }
-}
