+++ package.json
@@ -1,17 +1,21 @@
 {
   "name": "fictrail",
-  "version": "1.0.0",
-  "description": "A browser extension for viewing and searching your AO3 history",
+  "version": "0.0.0-dev",
+  "description": "A Tampermonkey userscript for viewing and searching your AO3 history",
   "scripts": {
-    "lint": "eslint *.js",
-    "lint:fix": "eslint *.js --fix"
+    "build": "node build.js",
+    "dev": "node build.js && echo 'Build complete - check dist/fictrail.user.js'",
+    "watch": "nodemon --watch src --exec 'npm run build'",
+    "lint": "eslint .",
+    "lint:fix": "eslint . --fix"
   },
   "devDependencies": {
     "@eslint/js": "^9.0.0",
     "eslint": "^9.35.0"
   },
   "keywords": [
-    "browser-extension",
+    "userscript",
+    "tampermonkey",
     "ao3",
     "fanfiction"
   ],
