+++ src/modules/constants.js
@@ -1,0 +1,2 @@
+// Constants
+const AO3_BASE_URL = 'https://archiveofourown.org';
